requirements.txt
PySide6
pandas
openpyxl
fpdf
mysql-connector-python
